core: "ipsec setup start" doesn't stick around
The command just starts the ipsec service and then exits, so if we
kill the VPN connection when it stops, we kill it every single time
and long before anything has happened.  Just don't do that.

Instead only fail if a child process exits with a non-zero exit
code, which probably does indicate an error.