core: rework connect process and subprocess handling
Instead of sprinkling sleep() calls around, waiting for processes to
be ready, actually wait until they are ready.  Since the ipsec
processes are only helpers that spawn the actual 'pluto' process,
we can simply wait until they exit to proceed.

Second, the user password handling was quite racy, often writing
the password long before the subprocess was ready.  Second, it didn't
work when running nm-openswan with --debug on a terminal, because
whack (the helper that asks for the password) uses getpass() which
opens /dev/tty instead of reading from stdin.  To handle that, we
must spawn the process that eventually asks for the password in
a pseudo-TTY with pipes so we can write the password to it when
it asks for one.