core: autodetect NEVER_DEFAULT by searching kernel IPSec routes
The IPSec daemon doesn't indicate to us whether upstream sent
routes (eg, CISCO_SPLIT_INC) so we have no way to easily determine
whether or not the VPN should get the default route automatically.

Instead, what we can do is ask the kernel for the IPSec security
association database, and look through the routes it returns for
an assocation between a subnet route and the VPN gateway we just
connected to.  This indicates that the gateway sent some subnets
that should be secured with IPSec, which in turn indicates that
the VPN should not receive the default route (because if it should,
why was the VPN sending specific subnets to secure?).