service: process the configuration in the service, not the helper
For multiple instance support we need to pass arguments to the helper. We use
the ifupdown option for that.

As we call our ifupdown script we need to manage the routes ourselves. That's
probably a good idea anyway. However, the helper never knows all the routes.
Therefore we have to move the logic of creating the IP4Config in the service.

The service adds a new bus object for the helper to talk to and the helper just
pushes in all the PLUTO_* environment variables to the service.

It's probably cleaner that way anyway + our DBus exposed VPN configuration now
makes sense. Yay!