auth-dialog: drop the .desktop file
It's generated incorrectly -- the @LIBEXECDIR@ is not substituted.

There's no point in keeping it; it was added because GNOME Shell needed
it to figure out the right window title; but it doesn't anymore and it
draws its own dialog using the external ui mode anyway.

It was dropped from NetworkManager-vpnc as well.