auth-dialog: drop the .desktop file
It's generated incorrectly -- the @LIBEXECDIR@ is not substituted.

There's no point in keeping it; it was added because GNOME Shell needed
it to figure out the right window title; but it doesn't anymore and it
draws its own dialog using the external ui mode anyway.

It was dropped from NetworkManager-vpnc as well.

(cherry picked from commit 9df39a1797e1405225679d59129890a6d98f317a)