properties: move nm_libreswan_config_read() to import_from_file()
nm_libreswan_config_read() had only one caller import_from_file() and is pretty simple.
It is not clear that it would ever be used outside of properties (because service only
writes the configuration, it doesn't read it).

Also, it had a a logging statement 'g_print ("Ignored line: '%s'", str)'
which however was never reached because the properties plugin never
sets @debug.