properties: fail import of files that have no "conn" section
Previously, import would only consider the lines that it understands,
ignoring any invalid line.

That means, a completely bogus file was still accepted and a
NMConnection witout ID was created (which later fails validation).

Do a minimum of validation and require a "conn" section.

Also, there could be multiple "conn" sections within a file.
Only consider the first and ignore the following.