properties,service: check for errors in write_config_option()
Check for errors during write, retry on EINTR and retry on
incomplete writes.

When write() or close() fails, report an error instead of logging
a warning with g_warning(). Using g_warning() is not great when run
as a plugin (because a runtime failure should not cause messages in
stdout). And is not great for service either, because it bypasses
whatever logging we want to perform there.

By reporting errors back to the caller, we no longer need to print
a g_warning() inside the utility function.