shared: refactor arguments for nm_libreswan_config_write()
Before, nm_libreswan_config_write() had fewer arguments, and based on those arguments
nm_libreswan_config_write() would decide what to do. For example, depending on the
presence of @bus_name, it would choose as connection name the ID or the
UUID.

Refactor the arguments, to move logic out of nm_libreswan_config_write():

 - instead of detecting the @con_name based on the presense of @bus_name,
   have an explict argument.
 - instead of deciding whether to append a @trailing_newline based on
   (openswan || !bus_name), let the caller decide.
 - the leftupdown_script is build based on @bus_name.

The reason for this change is, that with the next patch, we want to
pass more argument to the leftupdown script. Thus, we anyway would need
a to add a @leftupdown_script argument for that. But then keeping @bus_name
to control @con_name and @trailing_newline is confusing.