build: add configure option "--with-nm-ipsec-conf" and "--with-nm-ipsec-secrets-dir"
nm-libreswan passes "--config" to pluto, but it used as path SYSCONFDIR"/ipsec.conf".
Also, it would write the secret file to SYSCONFDIR"/ipsec.d/ipsec-%s.secrets".

However, SYSCONFDIR is "${prefix}/etc", so when installing nm-libreswan to
a different prefix, this is wrong.

Make these locations configurable, and change the defaults to
"/etc/ipsec.conf" and "/etc/ipsec.d".