service: always read stdout/stderr of ipsec command
We redirect stdout/stderr of the ipsec command to a pipe.

That is for one done so that we can print it as our own
logging output (and possibly disable it unless debug logging
is enabled).

It is also done, to make the command invocation silent, unless
debug logging is enabled.
Not redirecting unless debug logging is enabled is hence not an option,
because then our regular output would be spammed with debugging information.

However, as we create a pipe, we *must* read it and possible throw away
the data. Otherwise, the pipe might fill up and block the process which
wants to write to it.