all: add support to IKEv2
Add the brand new option 'ikev2', which maps the libreswan option of the
same name. When the option is found and is set with a value that enables
IKEv2, the option itself is passed as-is to libreswan while all the
XAUTH options are skipped (they are still added when the configuration is
IKEv1 based).
The configuration passed to libreswan will be of type IRAC (IPsec Remote
Access Client), enforcing CP mode. Anyway, no EAP method will be used as
it is not yet available in libreswan.