service: support detecting IPv6 SAD routes
Similarly to what we currently do for IPv4, also detect IPv6 SAD
routes based on XFRM policies.

We match entries where the selector's destination network is not /0
and the gateway (tmpl) is our VPN gateway. Note that the selector and
gateway can be of different address families; therefore we return a
separate array for IPv4 and IPv6 routes.