shared: make ipsec.conf parsing declarative
Reuse metadata used for producing ipsec.conf fragments to drive the
import of ipsec.conf fragments and ensure their soundness.

This makes import much more sensible: it now deals with quoting
properly, doesn't ignore errors silently. It also validates the handful
of properties we decided to be smart about and generate them, but not
allow the user to set them explicitly: it just checks whether what we'd
generate is the same as specified in input, and happily proceed if it
matches and fail otherwise.

Also, cover it with tests, like a real grownup.

https://gitlab.gnome.org/GNOME/NetworkManager-libreswan/-/merge_requests/38