shared: make ipsec.conf formatting declarative
At this point it is obvious that being clever about adding properties
that the user didn't specify was not the greatest idea.
nm_libreswan_get_ipsec_conf() has became a sad and miserable mess.
This attempts to untangle it.

The idea is to have a declarative description of the properties,
along with a routine that generates them if unspecified and flags that
specify how is it formatted.

This hopefully breaks the logic down into managable pieces while keeping
nm_libreswan_get_ipsec_conf() reasonably simple. The data, along with
the routine that produces a full (canonical, sanitized) configuration is
kept separate, because it could be used on import as well (a follow-up
commit), chopping of annother bit that has grown too complicated.

Moreover it does validation as well, so separate nm_ls_*_validate()
could be dropped off. Yes, ha ha ha, *yes*!