utils: fix export/import of cisco-unity
This is an odd property. We'd generating it if a made up vendor=Cisco
property was present and ignored if specified explicitly.

In an effort to make export/import more consistent, it ended up being
checked for and rejected if present. While an improvement over outright
ignoring it, it made connections we export not re-importable in turn.

Let's turn it into an ordinary non-synthetic property -- preserve it as
it is, and keep generating it if vendor=Cisco is present to keep
compatibility.