Support `leftsendcert` option
NetworkManager-libreswan does not validate values of this option, just
treat it as arbitrary string and send to libreswan daemon.

Instead of provide a list of accepted values, we allows arbitrary string
for this option because the valid value of this property might changed by
libreswan, and it is simpler for us trusting libreswan reply on invalid
argument.

Unit test case included.

Resolves: https://issues.redhat.com/browse/RHEL-104444

Signed-off-by: Gris Ge <fge@redhat.com>