service, editor: remove the "add" value for nm-connect-mode
In all cases, "nm-connect-mode=ondemand" is better than
"nm-connect-mode=add".

When bringing up two machines, the startup order is generally
unpredictable. "ondemand" on both sides resolves this; both peers
listen and they try to initiate the connection on outgoing traffic.

"ondemand" also provides better security. It installs the xfrm policy
immediately, ensuring that traffic never leaks unencrypted while
waiting for the tunnel to be established.

In "ondemand" mode, each peer tries to connect for 60 seconds on
outgoing traffic. If they fail or time out, they revert to a listening
state without marking the NetworkManager connection as failed.

If users still want to configure a client/server topology, they can
use "up" on the client and "ondemand" on the server ("ondemand" is a
superset of "add"). The advantage of this setup is that NetworkManager
reports an immediate error on the client in case of failure.

As a bonus, dropping "add" makes the implementation simpler, as we
don't need any changes in the NM daemon.